[plugins] Add init hook to plugins. Add ruby gem bin to path
