+++ plugins/ruby.json
@@ -6,5 +6,10 @@
     "RUBY_CONFDIR": "{{ .Virtenv }}",
     "GEMRC": "{{ .Virtenv }}/.gemrc",
     "GEM_HOME": "{{ .Virtenv }}"
+  },
+  "shell": {
+    "init_hook": [
+      "export PATH={{ .Virtenv }}/bin:$PATH"
+    ]
   }
 }
